+++ src/contentScript/messenger/strict/api.ts
@@ -67,3 +67,4 @@ export const insertButton = getMethod("INSERT_BUTTON");
 export const enableOverlay = getMethod("ENABLE_OVERLAY");
 export const disableOverlay = getMethod("DISABLE_OVERLAY");
 export const runMapArgs = getMethod("RUN_MAP_ARGS");
+export const getCopilotHostData = getMethod("GET_COPILOT_HOST_DATA");
