+++ src/contrib/automationanywhere/SetCopilotDataEffect.ts
@@ -19,7 +19,7 @@ import { validateRegistryId } from "@/types/helpers";
 import { type Schema } from "@/types/schemaTypes";
 import { type BrickArgs } from "@/types/runtimeTypes";
 import { EffectABC } from "@/types/bricks/effectTypes";
-import { setPartnerCopilotData } from "@/background/messenger/api";
+import { setPartnerCopilotData } from "@/background/messenger/strict/api";
 import { isLoadedInIframe } from "@/utils/iframeUtils";
 import { BusinessError } from "@/errors/businessErrors";
 import { propertiesToSchema } from "@/utils/schemaUtils";
